+++ clippy_lints/src/panic_unimplemented.rs
